backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.5
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
GeoAlchemy2>=0.14.3
anthropic>=0.84.0
groq>=0.4.0
python-dateutil>=2.8.0
Werkzeug>=3.0.0
marshmallow>=3.20.0
Pillow>=10.0.0
requests>=2.31.0
gunicorn>=21.0.0
pgvector>=0.2.0
sentence-transformers>=2.2.0
